Fix locating schema file problem
https://github.com/jmatsuzawa/gsettings-info/issues/10

Gsettings-info could not locate schema files whose name does not
correspond to the schema id. This commit change the way to locate file.

New method is just grep id=$SCHEMA *.gschema.xml. Unfortunately,
according to the gschema DTD [1], the attribute 'id' is not defined as
uniq. This method might not locate correctly a file. But it is very
rare in real case.

[1] https://git.gnome.org/browse/glib/plain/gio/gschema.dtd?h=glib-2-34